requirements.txt
anyio==3.6.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.89.1
Flask==2.2.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycodestyle==2.10.0
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
requests==2.28.2
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.14
uuid==1.30
uvicorn==0.20.0
Werkzeug==2.2.2
zipp==3.12.0